return int instead of float from tuiLinesGetCrossProduct
The lines are made up of int points so the cross product will always be an int too

+++ include/TUIC/line.h
@@ -120,7 +120,7 @@ int tuiLineGetDiagonalLength(const TuiLine line);
  *
  * @thread_safety This function does not access memory in ways that could cause a data race, so it is thread safe.
  */
-float tuiLinesGetCrossProduct(const TuiLine line_1, const TuiLine line_2);
+int tuiLinesGetCrossProduct(const TuiLine line_1, const TuiLine line_2);
 /*!
  * @brief Get if two @ref TuiLine are parallel.
  *
